backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlalchemy>=2.0.0
langchain>=0.3.0
langchain-openai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
